refactor copa
